Encode bytes into UTF-16 text string
Base64 is not space efficient. Base64 was intended to be a binary to
text encoding that would allow delivery of the data across various
mediums where the only commonly supported would be the most common
7-bit alpanumeric characters. Since the Base64 string is being stored
in a C#'s native string is composed of 16-bit characters encoded as
UTF-16, 9-bits worth of potential storage space is wasted for every
character used.

Each pair of bytes from the original data is taken to form the low and
high bytes of the new Unicode character that will be added to the
string.

If ever a pair will cause the character to fall within the surrogate
range (U+D800-U+DFFF), the low and high bytes are encoded as surrogate
pair of U+D800 + low, and U+DC00 + high.

The CompressionValue record type has been updated to show the UTF-16
string length. Previously it would store a Base64 string, but show the
byte lengths of the unencoded bytes.